Use a double parameter for "double expression"

+++ tck/src/main/java/ee/jakarta/tck/data/standalone/persistence/stateful/Inventory.java
@@ -59,5 +59,5 @@ public interface Inventory {
             ORDER BY price DESC, name ASC, productNum DESC
             """)
     List<Product> withDiscountedPriceUpTo(double max,
-                                          int percentOff);
+                                          double percentOff);
 }
